Split sound module
Create main_player and player modules
Move sounds population to window
Rename SoundObject to Sound
