+++ data/resources/grid-item.ui
@@ -5,19 +5,19 @@
       <object class="SoundItem">
 
         <binding name="playing">
-          <lookup name="playing" type="SoundObject">
+          <lookup name="playing" type="Sound">
             <lookup name="item">GtkListItem</lookup>
           </lookup>
         </binding>
 
         <binding name="title">
-          <lookup name="title" type="SoundObject">
+          <lookup name="title" type="Sound">
             <lookup name="item">GtkListItem</lookup>
           </lookup>
         </binding>
 
         <binding name="icon-name">
-          <lookup name="icon-name" type="SoundObject">
+          <lookup name="icon-name" type="Sound">
             <lookup name="item">GtkListItem</lookup>
           </lookup>
         </binding>
